release, debug 모두 정상 및 모두 릭도 안남음

+++ HEngine/HHH/Window/AppWindow.cpp
@@ -57,7 +57,7 @@ void AppWindow::onUpdate()
 	EngineKey::MouseX = -MouseForceX;
 	EngineKey::MouseY = -MouseForceY;
 	
-	GameEngine::get()->Update(TimeManager->Update());
+	GameEngine::get()->Update(static_cast<float>(TimeManager->Update()));
 }
 
 void AppWindow::onDestroy()
